feature (pa-20): isolated util functions and removed unused imports
