+++ backend/project_alpha/web/utils/send_email_utils.py
@@ -1,0 +1,8 @@
+def send_recovery_email(user):
+    '''
+    Sends personnalized password recovery e-mail to user.
+    P.S. Should be async.
+    TODO
+    '''
+    print("Send e-mail to " + str(user.email))
+
