+++ frontend/src/actions/authActions.ts
@@ -11,7 +11,7 @@ import {jwtDecode, IToken, autoRefresh } from '../services/TokenRefresh'
 import { MILLISECONDS_IN_SECOND, TEN_SECONDS_BEFORE_TOKEN_EXPIRE} from '../constants/valuableNumbers'
 import {errorMessage } from '../constants/errorAndSuccessMessages'
 import { achievementMessage } from '../constants/achievementMessages'
-import { StringLocale } from 'yup/lib/locale'
+import { emailFormatIsValid } from '../utils/FormatVerifications'
 
 export enum AuthActionType {
   LOG_OUT = 'auth/LOG_OUT',
@@ -113,10 +113,7 @@ export const userLogOut = () => (dispatch: Dispatch<logoutUserAction | CallHisto
     }
 }
 
-const validEmail = (email: string) => {
-  const re = /\S+@\S+\.\S+/
-  return re.test(email)
-}
+
 
 export const recover = (payload: IUserEmail) =>
   async (dispatch:Dispatch<logoutUserAction | CallHistoryMethodAction>): Promise<void> => {
@@ -124,7 +121,7 @@ export const recover = (payload: IUserEmail) =>
 
     try {
       const {email} = payload
-      if (validEmail(email)) {
+      if (emailFormatIsValid(email)) {
         const result = await authApiClient.recover(payload)
         const {message} = result
         toast.info(message)
