+++ frontend/src/utils/FormatVerifications.ts
@@ -1,0 +1,4 @@
+export const emailFormatIsValid = (email: string):boolean => {
+    const re = /\S+@\S+\.\S+/
+    return re.test(email)
+}
